+++ tests/test_config.py
@@ -1,11 +1,7 @@
 from inline_snapshot import snapshot
-from pydantic import SecretStr
 
 from kimi_cli.config import (
     Config,
-    LLMModel,
-    LLMProvider,
-    MoonshotSearchConfig,
     Services,
     get_default_config,
 )
@@ -15,18 +11,10 @@ def test_default_config():
     config = get_default_config()
     assert config == snapshot(
         Config(
-            default_model="kimi-k2-turbo-preview",
-            models={
-                "kimi-k2-turbo-preview": LLMModel(provider="kimi", model="kimi-k2-turbo-preview")
-            },
-            providers={
-                "kimi": LLMProvider(
-                    type="kimi",
-                    base_url="https://api.moonshot.cn/v1",
-                    api_key=SecretStr(""),
-                )
-            },
-            services=Services(moonshot_search=MoonshotSearchConfig(api_key=SecretStr(""))),
+            default_model="",
+            models={},
+            providers={},
+            services=Services(),
         )
     )
 
@@ -36,31 +24,14 @@ def test_default_config_dump():
     assert config.model_dump_json(indent=2, exclude_none=True) == snapshot(
         """\
 {
-  "default_model": "kimi-k2-turbo-preview",
-  "models": {
-    "kimi-k2-turbo-preview": {
-      "provider": "kimi",
-      "model": "kimi-k2-turbo-preview",
-      "max_context_size": 200000
-    }
-  },
-  "providers": {
-    "kimi": {
-      "type": "kimi",
-      "base_url": "https://api.moonshot.cn/v1",
-      "api_key": ""
-    }
-  },
+  "default_model": "",
+  "models": {},
+  "providers": {},
   "loop_control": {
     "max_steps_per_run": 100,
     "max_retries_per_step": 3
   },
-  "services": {
-    "moonshot_search": {
-      "base_url": "https://search.saas.moonshot.cn/v1/search",
-      "api_key": ""
-    }
-  }
+  "services": {}
 }\
 """
     )